requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.3
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.1
Django==3.2
django-allauth==0.44.0
django-crispy-forms==1.11.2
django-widget-tweaks==1.4.8
filelock==3.0.12
future==0.18.2
html5lib==1.1
idna==2.10
importlib-metadata==4.0.1
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==8.2.0
pycparser==2.20
PyJWT==2.0.1
PyPDF2==1.26.0
python-bidi==0.4.2
python3-openid==3.2.0
pytz==2021.1
reportlab==3.5.66
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.3.1
stripe==2.56.0
twilio==6.50.1
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.4.4
webencodings==0.5.1
xhtml2pdf==0.2.5
zipp==3.4.1
